1.加入配置文件可修改浮点数类型。2.修改readme。3.加入osqp-eigen源码编译。

+++ Mpc/Cpp/include/TinyMPC/types.hpp
@@ -12,7 +12,7 @@ using namespace Eigen;
 extern "C" {
 #endif
 
-    typedef double tinytype;  // should be double if you want to generate code
+    typedef float tinytype;  // should be double if you want to generate code
     typedef Eigen::Matrix<tinytype, Dynamic, Dynamic> tinyMatrix;
     typedef Eigen::Matrix<tinytype, Dynamic, 1> tinyVector;
